server/requirements.txt
openenv[core]>=0.2.0
fastapi>=0.115.0
uvicorn>=0.24.0
requests>=2.28.0
google-cloud-firestore>=2.19.0



